blueshield_doc_ai/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
black==23.12.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
Flask==2.3.3
frozenlist==1.4.1
functions-framework==3.5.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-documentai==2.21.1
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-oauth2-tool==0.0.3
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauth2client==4.1.3
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
platformdirs==4.1.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
yarl==1.9.4
